Various API changes
- Add default constructor for File()
- Add move constructor and assignment operator
- Initial implementation of file locking when file is open in write mode

Rationale:
Not supporting default constructions and/or move construction makes the
API awkard to use, e.g. it is not possible to declare an
optional/smart_ptr to File.

Adding the move constructor and assignment operator without safe-guards
opens up the following issue:

```c++
auto f = File::create_new_file("file.cool", ...); // OK
...
...
f = File::open_read_only("file.cool", ...); // File construction fails
```

Cooler validation during construction fails, because not all the
data/attributes may have yet been written to disk. Unfortunately this
leads to counter-intuitive error messages.

To workaround this issue, when creating a new Cooler, coolerpp will
create the `format-version` and `created-by` attributes immediately.
The `format-version` attribute is set to a sentinel value (currently 255)
during construction, and set to the final value (currently 3) last
thing when finalize() is called by the destructor.

Validation methods (which are called during File construction) check if
a Cooler was created by coolerpp, and if `format-version` is equal to 255.
When this is the case, coolerpp raises an exception that clearly states
that the Cooler was not (properly) closed, and that it is thus likely
corrupted or incomplete.