Bugfix
Fix segfault in move ctor/assignment operator.
Segfault was due to Index ctor, which takes a const reference to the bin
table and stores a raw pointer to the table in one of its member
variables.
The pointer to the bin table is of course invalidated after move,
leading to segfaults and other weird issues.

Index an bin table are now stored as unique_ptr in the File object.

Note for developers: prefer accessing the bin table and index through
the bin() and index() methods, as these check that the unique_ptr is not
null before dereferencing it.