Rework caching
Most datasets are read once upon file opening or reading weights.
Thus, it doesn't make sense to keep a large cache for those.
Furthermore, the only datasets that we may access somewhat randomly are
datases from the pixel table.
This commit updates several constants and a handful of functions to set
dataset access properties more wisely.

The commit also replaces the table of prime numbers with a function to
find the closest prime number given a number n.
For numbers smaller than 2**32, this usually takes up to 10-15ms (and more often 1-2ms).
This should lead to smaller binaries and also better performance when
using larger caches.