Add optimization for PixelSelector<N>::iterator operator++(int)
Without this optimization, looping over pixels using it++ shows very
poor performance.
Once exhausting the first chunk of pixels, calling it++ causes a lot of
unnecessary reads from disk.
This can be avoided by "refreshing" the buffer when there's still one pixel
available, so that the buffer is refreshed when calling it++, and also
reused in subsequent calls to ++it.