Simplify implementation of PixelSelector<N>::iterator
Initially, our intention was to implement PixelSelector<N>::iterator as
a bidirectional iterator. However this would introduce quite a bit of
complexity (mostly due to trans queries) in exchange for slightly more
flexibility. Thus, we decided to stick with the forward iterator
implementation we currently have. This means that the logic to detect
iterators at end can be simplified, as there's no need to know this
location ahead of time: we can simply check when the iterator stops
overlapping queries in both dimensions.